requirements.txt

mail-validator>=2.2.0
flask-wtf>=1.2.2
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
flask-login>=0.6.3
flask-mail>=0.10.0
requests>=2.32.3
sqlalchemy>=2.0.39
werkzeug>=3.1.3
wtforms>=3.2.1
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.164.0
oauthlib>=3.2.2
google-auth>=2.38.0
google-auth-httplib2>=0.2.0
authlib>=1.5.1
flask-migrate>=4.1.0
mysql-connector-python>=9.2.0
pymysql>=1.1.1
python-dotenv
